Moved App.tsx to Components
