+++ src/Components/App.tsx
@@ -2,11 +2,11 @@ import React from "react";
 import logo from "../assets/logo.svg";
 import "../css/App.css";
 import { useTheme } from "@mui/material/styles";
-import ThemeToggleButton from "../Components/ThemeToggleButton";
+import ThemeToggleButton from "./ThemeToggleButton";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Login from "../Components/Login";
-import forgotpwd from "./forgot_pwd";
-import ForgotPwd from "./forgot_pwd";
+import Login from "./Login";
+import forgotpwd from "../pages/forgot_pwd";
+import ForgotPwd from "../pages/forgot_pwd";
 
 function App() {
   const theme = useTheme();
